+++ src/pydvl/utils/config.py
@@ -13,16 +13,22 @@ __all__ = ["ParallelConfig", "MemcachedClientConfig", "MemcachedConfig"]
 class ParallelConfig:
     """Configuration for parallel computation backend.
 
-    :param backend: Type of backend to use. For now only 'ray' is supported.
+    :param backend: Type of backend to use.
+        Defaults to 'ray'
     :param address: Address of existing remote or local cluster to use.
-    :param n_local_workers: Number of workers (CPUs) to use when using a local ray cluster.
+    :param n_cpus_local: Number of CPUs to use when creating a local ray cluster.
+        This has no effect when using an existing ray cluster.
     :param logging_level: Logging level for the parallel backend's worker.
     """
 
     backend: Literal["sequential", "ray"] = "ray"
     address: Optional[Union[str, Tuple[str, int]]] = None
-    n_local_workers: Optional[int] = None
+    n_cpus_local: Optional[int] = None
     logging_level: int = logging.WARNING
+
+    def __post_init__(self) -> None:
+        if self.address is not None and self.n_cpus_local is not None:
+            raise ValueError("When `address` is set, `n_cpus_local` should be None.")
 
 
 @dataclass